Getting the external build server going end to end

+++ MonoProgram.Package/Utils/SftpExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
 using Renci.SshNet;
 
 namespace MonoProgram.Package.Utils
@@ -14,22 +15,37 @@ namespace MonoProgram.Package.Utils
         {
             foreach (var file in client.ListDirectory(path))
             {
+                if (file.Name == "." || file.Name == "..")
+                    continue;
+
                 if (file.IsDirectory)
-                {
                     client.Clear(file.Name);
-                }
                 file.Delete();
             }
         }
 
-        public static void Upload(this SftpClient client, string localFolder)
+        public static void Upload(this SftpClient client, string basePath, string path, HashSet<string> createdDirectories)
         {
-            var directory = new DirectoryInfo(localFolder);
-            foreach (var file in directory.EnumerateFiles())
+            var fullPath = Path.Combine(basePath, path);
+            if (Directory.Exists(fullPath))
             {
-                using (var fileStream = file.OpenRead())
+                var directory = new DirectoryInfo(path);
+                foreach (var file in directory.EnumerateFiles())
                 {
-                    client.UploadFile(fileStream, file.Name);
+                    using (var fileStream = file.OpenRead())
+                    {
+                        client.UploadFile(fileStream, file.Name);
+                    }
+                }
+            }
+            else
+            {
+                var directory = Path.GetDirectoryName(path);
+                if (createdDirectories.Add(directory))
+                    client.CreateFullDirectory(directory);
+                using (var fileStream = new FileInfo(fullPath).OpenRead())
+                {
+                    client.UploadFile(fileStream, path);
                 }
             }
         }
@@ -40,7 +56,7 @@ namespace MonoProgram.Package.Utils
             var currentPath = ".";
             foreach (var part in parts)
             {
-                var newPath = Path.Combine(currentPath, part);
+                var newPath = currentPath + "/" + part;
                 if (!client.Exists(newPath)) 
                     client.CreateDirectory(newPath);
                 currentPath = newPath;
